proyectocc-backend/requirements.txt
ffcuesplitter
pytest
Flask
Werkzeug
requests
flask-cors
psycopg2-binary
shutil